Added all logic and CSS to Create a Review button

+++ backend/db/models/review.js
@@ -27,7 +27,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     stars: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(2,1),
       allowNull: false,
       validate: {
         len: [1, 5]
